+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 🐞 Fixes
 
+- Fixed an issue with `moon setup` failing when proto is not available.
 - Fixed task input inferrence to not include the `$HOME`, `$USER`, and `$PWD` environment variables.
 - Fixed some arg/shell quoting issues.